==> 2_siswa_klasifikasi.ipynb <==
import pandas as pd
import numpy as np

# 1. KLASIFIKASI SKOR TOTAL SISWA

# Load data siswa
df_siswa = pd.read_csv("files/data_CTnorm_siswa.csv", sep=';')
print("\nJumlah data siswa:", len(df_siswa))

print("\nData Siswa:")
df_siswa[["Nama", "Kategori", "Nilai", "CT_norm"]].head() 

# 2. KLASIFIKASI SKOR TOTAL

# Cek Trisection 
q33 = df_siswa["CT_norm"].quantile(0.33).round(3)
q66 = df_siswa["CT_norm"].quantile(0.66).round(3)

print("\nCek Trisection CT_norm Siswa:")
print("\nNilai CT_norm pada kuartil 33%  :", q33)
print("Nilai CT_norm pada kuartil 66%  :", q66)

# Klasifikasi siswa
def klasifikasi_ct(x):
    if x <= q33:
        return 'Rendah'
    elif x <= q66:
        return 'Sedang'
    else:
        return 'Tinggi'

df_siswa['Level_CT'] = df_siswa['CT_norm'].apply(klasifikasi_ct)

print("\nKlasifikasi Level CT Siswa:")
df_siswa[['Nama', 'CT_norm', 'Level_CT']].head()


# 3. JUMLAH SISWA PER KLASIFIKASI LEVEL CT

# Hitung siswa per level CT
jumlah_siswa = df_siswa['Level_CT'].value_counts().sort_index()
persentase = (jumlah_siswa / jumlah_siswa.sum() * 100).round(2)

tabel_ct = pd.DataFrame({
    'Level': jumlah_siswa.index,
    'Jumlah Siswa': jumlah_siswa.values,
    'Persentase (%)': persentase.values
})

print("\nTabel Distribusi Tingkat Literasi Digital Siswa")
tabel_ct


# CEK OUTPUT DATA SISWA
df_siswa[[
    "Id",
    "Nama",
    "SekolahNama",
    "Kelas",
    "Kategori",
    "Nilai",
    "CT_norm",
    "Level_CT"
]].head()

# pilih kolom yang mau disimpan
df_save_siswa = df_siswa[[
    "Id",
    "Nama",
    "SekolahNama",
    "Kelas",
    "Kategori",
    "Nilai",
    "CT_norm",
    "Level_CT"
]]

# simpan ke CSV
df_save_siswa.to_csv(
    "files/data_siswa_klasifikasi.csv",
    index=False,
    float_format="%.3f"
)

print("\nData siswa berhasil disimpan")
